Reorgnize class functions
