+++ code-env/python/spec/requirements.txt
@@ -1,1 +1,76 @@
-spacy==3.0
+appdirs==1.4.4
+astroid==2.5.1
+black==20.8b1
+blis==0.7.4
+catalogue==2.0.1
+certifi==2020.12.5
+chardet==3.0.4
+click==7.1.2
+contextvars==2.4
+cymem==2.0.5
+dataclasses==0.8
+DAWG-Python==0.7.2
+docopt==0.6.2
+emoji==0.6.0
+fastcore==1.3.1
+flake8==3.8.4
+idna==2.8
+immutables==0.15
+importlib-metadata==3.7.0
+isort==5.7.0
+jieba==0.42.1
+Jinja2==2.11.3
+joblib==1.0.1
+langid==1.1.6
+lazy-object-proxy==1.5.2
+MarkupSafe==1.1.1
+mccabe==0.6.1
+murmurhash==1.0.5
+mypy-extensions==0.4.3
+numpy==1.19.5
+packaging==20.9
+pandas==0.23.4
+pathspec==0.8.1
+pathy==0.4.0
+plac==1.1.3
+preshed==3.0.5
+pycld3==0.20
+pycodestyle==2.6.0
+pydantic==1.7.3
+pyflakes==2.2.0
+pylint==2.7.2
+pymorphy2==0.9.1
+pymorphy2-dicts-ru==2.4.417127.4579844
+pymorphy2-dicts-uk==2.4.1.1.1460299261
+pyparsing==2.4.7
+pythainlp==2.2.6
+python-crfsuite==0.9.7
+python-dateutil==2.8.0
+pytz==2019.2
+pyvi==0.1
+regex==2020.10.28
+requests==2.22.0
+scikit-learn==0.24.1
+scipy==1.5.4
+six==1.15.0
+sklearn-crfsuite==0.3.6
+smart-open==3.0.0
+spacy==3.0.0
+spacy-legacy==3.0.1
+spacy-lookups-data==0.3.2
+spacymoji==2.0.0
+srsly==2.4.0
+symspellpy==6.7.0
+tabulate==0.8.9
+thinc==8.0.2
+threadpoolctl==2.1.0
+tinydb==4.4.0
+toml==0.10.2
+tqdm==4.50.2
+typed-ast==1.4.2
+typer==0.3.2
+typing-extensions==3.7.4.3
+urllib3==1.25.11
+wasabi==0.8.2
+wrapt==1.12.1
+zipp==3.4.0
